feat: Mejoras en logica y correccion de errores

+++ src/main/java/com/ansicode/SistemaAdministracionGym/pago/PagoRepository.java
@@ -8,7 +8,7 @@ import org.springframework.data.domain.Pageable;
 public interface PagoRepository extends JpaRepository<Pago, Long> {
 
     Page<Pago> findByMembresiaClienteId(
-            Integer membresiaClienteId,
+            Long membresiaClienteId,
             Pageable pageable
     );
